+++ src/providers/openai/index.ts
@@ -1,6 +1,5 @@
 import type { Cache } from 'cache-manager';
 import OpenAI from 'openai';
-import type { Metadata } from 'openai/resources/shared';
 import { fetchWithCache, getCache, isCacheEnabled } from '../../cache';
 import { getEnvString, getEnvFloat, getEnvInt } from '../../envars';
 import logger from '../../logger';
@@ -19,9 +18,9 @@ import type { EnvOverrides } from '../../types/env';
 import { renderVarsInObject } from '../../util';
 import { maybeLoadFromExternalFile } from '../../util';
 import { safeJsonStringify } from '../../util/json';
-import { sleep } from '../../util/time';
 import { ellipsize } from '../../utils/text';
-import { calculateCost, REQUEST_TIMEOUT_MS, parseChatPrompt, toTitleCase } from '../shared';
+import { calculateCost, REQUEST_TIMEOUT_MS, parseChatPrompt } from '../shared';
+import type { OpenAiSharedOptions } from './types';
 import type { OpenAiFunction, OpenAiTool } from './util';
 
 // see https://platform.openai.com/docs/models
@@ -141,16 +140,6 @@ export const OPENAI_COMPLETION_MODELS = [
     id: 'text-babbage-002',
   },
 ];
-
-interface OpenAiSharedOptions {
-  apiKey?: string;
-  apiKeyEnvar?: string;
-  apiHost?: string;
-  apiBaseUrl?: string;
-  organization?: string;
-  cost?: number;
-  headers?: { [key: string]: string };
-}
 
 export type OpenAiCompletionOptions = OpenAiSharedOptions & {
   temperature?: number;
@@ -738,250 +727,6 @@ export class OpenAiChatCompletionProvider extends OpenAiGenericProvider {
   }
 }
 
-type OpenAiAssistantOptions = OpenAiSharedOptions & {
-  modelName?: string;
-  instructions?: string;
-  tools?: OpenAI.Beta.Threads.ThreadCreateAndRunParams['tools'];
-  /**
-   * If set, automatically call these functions when the assistant activates
-   * these function tools.
-   */
-  functionToolCallbacks?: Record<
-    OpenAI.FunctionDefinition['name'],
-    (arg: string) => Promise<string>
-  >;
-  metadata?: Metadata;
-  temperature?: number;
-  toolChoice?:
-    | 'none'
-    | 'auto'
-    | 'required'
-    | { type: 'function'; function?: { name: string } }
-    | { type: 'file_search' };
-  attachments?: OpenAI.Beta.Threads.Message.Attachment[];
-  tool_resources?: OpenAI.Beta.Threads.ThreadCreateAndRunParams['tool_resources'];
-};
-
-export class OpenAiAssistantProvider extends OpenAiGenericProvider {
-  assistantId: string;
-  assistantConfig: OpenAiAssistantOptions;
-
-  constructor(
-    assistantId: string,
-    options: { config?: OpenAiAssistantOptions; id?: string; env?: EnvOverrides } = {},
-  ) {
-    super(assistantId, options);
-    this.assistantConfig = options.config || {};
-    this.assistantId = assistantId;
-  }
-
-  async callApi(
-    prompt: string,
-    context?: CallApiContextParams,
-    callApiOptions?: CallApiOptionsParams,
-  ): Promise<ProviderResponse> {
-    if (!this.getApiKey()) {
-      throw new Error(
-        'OpenAI API key is not set. Set the OPENAI_API_KEY environment variable or add `apiKey` to the provider config.',
-      );
-    }
-
-    const openai = new OpenAI({
-      apiKey: this.getApiKey(),
-      organization: this.getOrganization(),
-      baseURL: this.getApiUrl(),
-      maxRetries: 3,
-      timeout: REQUEST_TIMEOUT_MS,
-      defaultHeaders: this.assistantConfig.headers,
-    });
-
-    const messages = parseChatPrompt(prompt, [
-      {
-        role: 'user',
-        content: prompt,
-        ...(this.assistantConfig.attachments
-          ? { attachments: this.assistantConfig.attachments }
-          : {}),
-      },
-    ]) as OpenAI.Beta.Threads.ThreadCreateParams.Message[];
-    const body: OpenAI.Beta.Threads.ThreadCreateAndRunParams = {
-      assistant_id: this.assistantId,
-      model: this.assistantConfig.modelName || undefined,
-      instructions: this.assistantConfig.instructions || undefined,
-      tools:
-        maybeLoadFromExternalFile(renderVarsInObject(this.assistantConfig.tools, context?.vars)) ||
-        undefined,
-      metadata: this.assistantConfig.metadata || undefined,
-      temperature: this.assistantConfig.temperature || undefined,
-      tool_choice: this.assistantConfig.toolChoice || undefined,
-      tool_resources: this.assistantConfig.tool_resources || undefined,
-      thread: {
-        messages,
-      },
-    };
-
-    logger.debug(`Calling OpenAI API, creating thread run: ${JSON.stringify(body)}`);
-    let run: OpenAI.Beta.Threads.Run;
-    try {
-      run = await openai.beta.threads.createAndRun(body);
-    } catch (err) {
-      return failApiCall(err);
-    }
-
-    logger.debug(`\tOpenAI thread run API response: ${JSON.stringify(run)}`);
-
-    while (true) {
-      const currentRun = await openai.beta.threads.runs.retrieve(run.thread_id, run.id);
-
-      if (currentRun.status === 'completed') {
-        run = currentRun;
-        break;
-      }
-
-      if (currentRun.status === 'requires_action') {
-        const requiredAction = currentRun.required_action;
-        if (requiredAction === null || requiredAction.type !== 'submit_tool_outputs') {
-          run = currentRun;
-          break;
-        }
-        const functionCallsWithCallbacks: OpenAI.Beta.Threads.Runs.RequiredActionFunctionToolCall[] =
-          requiredAction.submit_tool_outputs.tool_calls.filter((toolCall) => {
-            return (
-              toolCall.type === 'function' &&
-              toolCall.function.name in (this.assistantConfig.functionToolCallbacks ?? {})
-            );
-          });
-        if (functionCallsWithCallbacks.length === 0) {
-          run = currentRun;
-          break;
-        }
-        logger.debug(
-          `Calling functionToolCallbacks for functions: ${functionCallsWithCallbacks.map(
-            ({ function: { name } }) => name,
-          )}`,
-        );
-        const toolOutputs = await Promise.all(
-          functionCallsWithCallbacks.map(async (toolCall) => {
-            logger.debug(
-              `Calling functionToolCallbacks[${toolCall.function.name}]('${toolCall.function.arguments}')`,
-            );
-            const functionResult = await this.assistantConfig.functionToolCallbacks![
-              toolCall.function.name
-            ](toolCall.function.arguments);
-            return {
-              tool_call_id: toolCall.id,
-              output: functionResult,
-            };
-          }),
-        );
-        logger.debug(
-          `Calling OpenAI API, submitting tool outputs for ${currentRun.thread_id}: ${JSON.stringify(
-            toolOutputs,
-          )}`,
-        );
-        try {
-          run = await openai.beta.threads.runs.submitToolOutputs(
-            currentRun.thread_id,
-            currentRun.id,
-            {
-              tool_outputs: toolOutputs,
-            },
-          );
-        } catch (err) {
-          return failApiCall(err);
-        }
-        continue;
-      }
-
-      if (
-        currentRun.status === 'failed' ||
-        currentRun.status === 'cancelled' ||
-        currentRun.status === 'expired'
-      ) {
-        run = currentRun;
-        break;
-      }
-
-      await sleep(1000);
-    }
-
-    if (run.status !== 'completed' && run.status !== 'requires_action') {
-      if (run.last_error) {
-        return {
-          error: `Thread run failed: ${run.last_error.message}`,
-        };
-      }
-      return {
-        error: `Thread run failed: ${run.status}`,
-      };
-    }
-
-    // Get run steps
-    logger.debug(`Calling OpenAI API, getting thread run steps for ${run.thread_id}`);
-    let steps;
-    try {
-      steps = await openai.beta.threads.runs.steps.list(run.thread_id, run.id, {
-        order: 'asc',
-      });
-    } catch (err) {
-      return failApiCall(err);
-    }
-    logger.debug(`\tOpenAI thread run steps API response: ${JSON.stringify(steps)}`);
-
-    const outputBlocks = [];
-    for (const step of steps.data) {
-      if (step.step_details.type === 'message_creation') {
-        logger.debug(`Calling OpenAI API, getting message ${step.id}`);
-        let message;
-        try {
-          message = await openai.beta.threads.messages.retrieve(
-            run.thread_id,
-            step.step_details.message_creation.message_id,
-          );
-        } catch (err) {
-          return failApiCall(err);
-        }
-        logger.debug(`\tOpenAI thread run step message API response: ${JSON.stringify(message)}`);
-
-        const content = message.content
-          .map((content) =>
-            content.type === 'text' ? content.text.value : `<${content.type} output>`,
-          )
-          .join('\n');
-        outputBlocks.push(`[${toTitleCase(message.role)}] ${content}`);
-      } else if (step.step_details.type === 'tool_calls') {
-        for (const toolCall of step.step_details.tool_calls) {
-          if (toolCall.type === 'function') {
-            outputBlocks.push(
-              `[Call function ${toolCall.function.name} with arguments ${toolCall.function.arguments}]`,
-            );
-            outputBlocks.push(`[Function output: ${toolCall.function.output}]`);
-          } else if (toolCall.type === 'file_search') {
-            outputBlocks.push(`[Ran file search]`);
-          } else if (toolCall.type === 'code_interpreter') {
-            const output = toolCall.code_interpreter.outputs
-              .map((output) => (output.type === 'logs' ? output.logs : `<${output.type} output>`))
-              .join('\n');
-            outputBlocks.push(`[Code interpreter input]`);
-            outputBlocks.push(toolCall.code_interpreter.input);
-            outputBlocks.push(`[Code interpreter output]`);
-            outputBlocks.push(output);
-          } else {
-            outputBlocks.push(`[Unknown tool call type: ${(toolCall as any).type}]`);
-          }
-        }
-      } else {
-        outputBlocks.push(`[Unknown step type: ${(step.step_details as any).type}]`);
-      }
-    }
-
-    return {
-      output: outputBlocks.join('\n\n').trim(),
-      tokenUsage: getTokenUsage(run, false),
-    };
-  }
-}
-
 type OpenAiImageOptions = OpenAiSharedOptions & {
   size?: string;
 };
@@ -1178,3 +923,8 @@ export const DefaultGradingJsonProvider = new OpenAiChatCompletionProvider('gpt-
 });
 export const DefaultSuggestionsProvider = new OpenAiChatCompletionProvider('gpt-4o-2024-05-13');
 export const DefaultModerationProvider = new OpenAiModerationProvider('omni-moderation-latest');
+
+export { OpenAiAssistantProvider } from './assistant';
+export type { OpenAiAssistantOptions } from './assistant';
+export type { OpenAiSharedOptions } from './types';
+export type { OpenAiFunction, OpenAiTool } from './util';
