+++ src/providers/openai/types.ts
@@ -1,0 +1,9 @@
+export interface OpenAiSharedOptions {
+  apiKey?: string;
+  apiKeyEnvar?: string;
+  apiHost?: string;
+  apiBaseUrl?: string;
+  organization?: string;
+  cost?: number;
+  headers?: { [key: string]: string };
+}
